+++ x/bank/types/params_legacy.go
@@ -1,0 +1,23 @@
+package types
+
+import paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
+
+var (
+	// KeySendEnabled is store's key for SendEnabled Params
+	// DEPRECATED: Use the SendEnabled functionality in the keeper.
+	KeySendEnabled = []byte("SendEnabled")
+	// KeyDefaultSendEnabled is store's key for the DefaultSendEnabled option
+	KeyDefaultSendEnabled = []byte("DefaultSendEnabled")
+)
+
+// DEPRECATED: ParamKeyTable for bank module.
+func ParamKeyTable() paramtypes.KeyTable {
+	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
+}
+
+// DEPRECATED: ParamSetPairs implements params.ParamSet
+func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
+	return paramtypes.ParamSetPairs{
+		paramtypes.NewParamSetPair(KeyDefaultSendEnabled, &p.DefaultSendEnabled, validateIsBool),
+	}
+}